+++ src/error-boundary.js
@@ -2,11 +2,12 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import Logger from './logger/logger';
 import './error-boundary.scss';
+import fallback from './assets/fallback.png';
 
 class ErrorBoundary extends Component {
   constructor(props) {
     super(props);
-    this.state = { hasError: false };
+    this.state = { hasError: false, errorMessage: null, errorStackTrace: null };
   }
 
   // Update state so the next render will show the fallback UI.
@@ -19,14 +20,29 @@ class ErrorBoundary extends Component {
 
     const { errorHandler } = this.props;
     errorHandler(error, errorInfo);
+
+    this.setState({
+      errorMessage: error.message.toString(),
+      errorStackTrace: errorInfo.componentStack,
+    });
   }
 
   render() {
-    const { hasError } = this.state;
+    const { hasError, errorMessage, errorStackTrace } = this.state;
+
     if (hasError) {
       return (
-        <div className="error-boundary">
-          <div className="error-boundary-loader">Loading...</div>
+        <div
+          className="error-boundary"
+          style={{ backgroundImage: `url(${fallback})` }}
+        >
+          <div className="error-boundary-box">
+            <div className="error-boundary-header">Seneste log hændelser</div>
+            <pre className="error-boundary-stacktrace">
+              {errorMessage}
+              {errorStackTrace}
+            </pre>
+          </div>
         </div>
       );
     }